+++ src/control/directive.ts
@@ -18,6 +18,7 @@ import { Subscription } from 'rxjs/Subscription';
 import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/distinctUntilChanged';
+import 'rxjs/add/operator/do';
 
 import {
   SetValueAction,
@@ -28,6 +29,7 @@ import {
 } from '../actions';
 import { FormControlState, FormControlValueTypes } from '../state';
 import { selectValueAccessor } from '../value-accessors';
+import { NgrxValueConverter, NgrxValueConverters } from './value-converter';
 
 const CHANGE = 'change';
 const BLUR = 'blur';
@@ -35,24 +37,26 @@ const BLUR = 'blur';
 @Directive({
   selector: '[ngrxFormControlState]',
 })
-export class NgrxFormControlDirective<TValue extends FormControlValueTypes> implements AfterViewInit, OnDestroy {
-  @Input() set ngrxFormControlState(newState: FormControlState<TValue>) {
+export class NgrxFormControlDirective<TStateValue extends FormControlValueTypes, TViewValue = TStateValue> implements AfterViewInit, OnDestroy {
+  @Input() set ngrxFormControlState(newState: FormControlState<TStateValue>) {
     if (!newState) {
       throw new Error('The control state must not be undefined!');
     }
 
     if (this.state && newState.id !== this.state.id) {
-      this.viewValueIsKnown = false;
-      this.viewValue = undefined as any;
+      this.stateValue = newState.value;
+      const viewValue = this.ngrxValueConverter.convertStateToViewValue(newState.value);
+      this.valueAccessor.writeValue(viewValue);
     }
 
     this.state = newState;
     this.stateSubject$.next(newState);
   }
 
-  @Input() ngrxUpdateOn: typeof CHANGE | typeof BLUR = CHANGE;
+  @Input() ngrxUpdateOn: 'change' | 'blur' = CHANGE;
   @Input() ngrxEnableFocusTracking = false;
   @Input() ngrxEnableLastKeydownCodeTracking = false;
+  @Input() ngrxValueConverter: NgrxValueConverter<TViewValue, TStateValue> = NgrxValueConverters.identity<any>();
 
   // TODO: move this into a separate directive
   // automatically apply the attribute that's used by the CDK to set initial focus
@@ -62,22 +66,22 @@ export class NgrxFormControlDirective<TValue extends FormControlValueTypes> impl
 
   private subscriptions: Subscription[] = [];
 
-  private state: FormControlState<TValue>;
-  private stateSubject$ = new BehaviorSubject<FormControlState<TValue>>(this.state);
+  private state: FormControlState<TStateValue>;
+  private stateSubject$ = new BehaviorSubject<FormControlState<TStateValue>>(this.state);
   private valueAccessor: ControlValueAccessor;
 
   // the <any> cast is required due to a mismatch in the typing of lift() between Observable and BehaviorSubject
-  private get state$(): Observable<FormControlState<TValue>> { return this.stateSubject$ as any; }
+  private get state$(): Observable<FormControlState<TStateValue>> { return this.stateSubject$ as any; }
 
-  // we have to store the last reported value since when the action to update the state
-  // is dispatched a new state will be received inside the directive, which in turn would
-  // trigger a view update; however, most input elements move the cursor to the end of the
-  // input when a new value is written programmatically which means whenever the user
-  // types something the cursor is forced to the end of the input; to prevent this
-  // behavior we compare the last reported value with the value to be set and filter out
-  // those values that are equal to the last reported value
-  private viewValueIsKnown: boolean;
-  private viewValue: TValue;
+  // we have to store the latest known state value since most input elements don't play nicely with
+  // setting the same value again (e.g. input elements move the cursor to the end of the input when
+  // a new value is set which means whenever the user types something the cursor is forced to the
+  // end of the input) which would for example happen every time a new value is pushed to the state
+  // since when the action to update the state is dispatched a new state will be received inside
+  // the directive, which in turn would trigger a view update; to prevent this behavior we compare
+  // the latest known state value with the value to be set and filter out those values that are equal
+  // to the latest known value
+  private stateValue: TStateValue;
 
   constructor(
     private el: ElementRef,
@@ -88,21 +92,16 @@ export class NgrxFormControlDirective<TValue extends FormControlValueTypes> impl
     this.valueAccessor = selectValueAccessor(valueAccessors);
   }
 
-  @Input() convertViewValue: (value: any) => TValue = value => value;
-  @Input() convertModelValue: (value: TValue) => any = value => value;
-
   ngAfterViewInit() {
     if (!this.state) {
       throw new Error('The form state must not be undefined!');
     }
 
-    this.valueAccessor.registerOnChange((newValue: TValue) => {
-      newValue = this.convertViewValue(newValue);
-      this.viewValueIsKnown = true;
-      this.viewValue = newValue;
+    this.valueAccessor.registerOnChange((viewValue: TViewValue) => {
+      this.stateValue = this.ngrxValueConverter.convertViewToStateValue(viewValue);
 
-      if (this.viewValue !== this.state.value && this.ngrxUpdateOn === CHANGE) {
-        this.actionsSubject.next(new SetValueAction(this.state.id, this.viewValue));
+      if (this.stateValue !== this.state.value && this.ngrxUpdateOn === CHANGE) {
+        this.actionsSubject.next(new SetValueAction(this.state.id, this.stateValue));
       }
     });
 
@@ -111,20 +110,18 @@ export class NgrxFormControlDirective<TValue extends FormControlValueTypes> impl
         this.actionsSubject.next(new MarkAsTouchedAction(this.state.id));
       }
 
-      if (this.viewValue !== this.state.value && this.ngrxUpdateOn === BLUR) {
-        this.actionsSubject.next(new SetValueAction(this.state.id, this.viewValue));
+      if (this.stateValue !== this.state.value && this.ngrxUpdateOn === BLUR) {
+        this.actionsSubject.next(new SetValueAction(this.state.id, this.stateValue));
       }
     });
 
     this.subscriptions.push(
       this.state$
-        .map(s => this.convertModelValue(s.value))
-        .filter(v => !this.viewValueIsKnown || v !== this.viewValue)
-        .subscribe(value => {
-          this.valueAccessor.writeValue(value);
-          this.viewValueIsKnown = true;
-          this.viewValue = value;
-        })
+        .map(s => s.value)
+        .filter(v => v !== this.stateValue)
+        .do(value => this.stateValue = value)
+        .map(value => this.ngrxValueConverter.convertStateToViewValue(value))
+        .subscribe(value => this.valueAccessor.writeValue(value))
     );
 
     if (this.valueAccessor.setDisabledState) {