+++ src/ngrx-forms.ts
@@ -3,6 +3,7 @@ export * from './state';
 export { formControlReducer } from './control/reducer';
 export { formGroupReducer } from './group/reducer';
 export { NgrxFormControlDirective } from './control/directive';
+export { NgrxValueConverter, NgrxValueConverters } from './control/value-converter';
 export { NgrxFormDirective } from './group/directive';
 export * from './update-functions';